Move candlechart components to different directory
