+++ src/components/charts/CandleChart.vue
