+++ src/components/charts/PlotConfigurator.vue
